Minimize duplication between broker-service and brokerd
