+++ tools/brokerd/brokerd.hpp
@@ -1,0 +1,5 @@
+namespace pstore {
+    namespace broker {
+        int run_broker (switches opt);
+    }
+} // namespace pstore
